+++ packages/plugin-connector/src/plugin-discovery/load-config.spec.js
@@ -1,10 +1,10 @@
 import { vi, describe, it, expect } from 'vitest';
 
 vi.mock('fs/promises');
+vi.mock('fs');
 
-import fs from 'fs/promises';
+import fs from 'fs';
 import {
-	defaultConfig,
 	handleAt,
 	handleAtParsed,
 	invalidMinimalConfig,
@@ -15,43 +15,45 @@ import {
 } from './load-config.fixture';
 import { loadConfig } from './load-config';
 
-fs.readFile = /** @type {any} */ (vi.fn());
-const mockedReadFile = vi.mocked(fs.readFile);
+fs.readFileSync = /** @type {any} */ (vi.fn());
+const mockedReadFileSync = vi.mocked(fs.readFileSync);
 
 describe('loadConfig', () => {
 	it('should load a valid configuration', async () => {
-		mockedReadFile.mockResolvedValueOnce(validMinimalConfig);
+		mockedReadFileSync.mockImplementationOnce(() => {
+			return validMinimalConfig;
+		});
 
-		const config = await loadConfig(__dirname);
+		const config = loadConfig(__dirname);
 
 		expect(config).toEqual(validMinimalConfigParsed);
 	});
 
 	it('should properly escape @ symbols in keys when appropriate', async () => {
-		mockedReadFile.mockResolvedValueOnce(handleAt);
+		mockedReadFileSync.mockReturnValueOnce(handleAt);
 
-		const config = await loadConfig(__dirname);
+		const config = loadConfig(__dirname);
 
 		expect(config).toEqual(handleAtParsed);
 	});
 
 	it('should load a verbose configuration', async () => {
-		mockedReadFile.mockResolvedValueOnce(validConfig);
+		mockedReadFileSync.mockReturnValueOnce(validConfig);
 
-		const config = await loadConfig(__dirname);
+		const config = loadConfig(__dirname);
 
 		expect(config).toEqual(validConfigParsed);
 	});
 
-	it('should fail to load invalid configuration, but recover safely', async () => {
-		mockedReadFile.mockResolvedValueOnce(invalidMinimalConfig);
-		const config = loadConfig(__dirname);
-		expect(config).rejects.toBeInstanceOf(Error);
+	it('should fail to load invalid configuration', async () => {
+		mockedReadFileSync.mockReturnValueOnce(invalidMinimalConfig);
+		expect(() => loadConfig(__dirname)).toThrowError();
 	});
 
-	it('should fail to load missing configuration, but recover safely', async () => {
-		mockedReadFile.mockRejectedValueOnce(new Error('ENOENT'));
-		const config = await loadConfig(__dirname);
-		expect(config).toEqual({ components: defaultConfig, databases: {} });
+	it('should fail to load missing configuration', async () => {
+		mockedReadFileSync.mockImplementationOnce(() => {
+			throw new Error('ENOENT');
+		});
+		expect(() => loadConfig(__dirname)).toThrowError();
 	});
 });