+++ sites/example-project/src/pages/+layout.svelte
@@ -8,8 +8,7 @@
 	let fileTree = {
 		label: 'Home',
 		href: '/',
-		children: {},
-		isTemplated: false
+		children: {}
 	};
 	pagePaths.forEach(function (path) {
 		path.split('/').reduce(function (r, e) {
@@ -18,14 +17,12 @@
 				return (r['href'] = href);
 			} else {
 				let label = e.includes('[') ? undefined : e.replace(/_/g, ' ').replace(/-/g, ' ');
-				r.isTemplated = e.includes('[');
 				return (
 					r?.children[e] ||
 					(r.children[e] = {
 						label,
 						children: {},
-						href: undefined,
-						isTemplated: false
+						href: undefined
 					})
 				);
 			}
@@ -62,175 +59,56 @@
 </script>
 
 <script>
+	import '@evidence-dev/tailwind/fonts.css';
 	import '../app.css';
-	import { navigating } from '$app/stores';
-	import { blur } from 'svelte/transition';
+	import { navigating, page } from '$app/stores';
 	import { dev } from '$app/environment';
-
-	export let data;
-
-	let open = false;
-	//TODO: Offer this as a build parameter
-	// in dev. mode prevent prefetch on "hover"
+	import {
+		LoadingSkeleton,
+		Sidebar,
+		BreadCrumbs,
+		Header,
+		TableOfContents
+	} from '@evidence-dev/core-components';
 	const prefetchStrategy = dev ? 'tap' : 'hover';
+
+	let mobileSidebarOpen = false;
+
+	$: if ($navigating) {
+		mobileSidebarOpen = false;
+	}
 </script>
 
-<!-- eslint-disable no-undef -->
-
-{#if $navigating}
-	<LoadingIndicator />
-{/if}
-
-<div data-sveltekit-preload-data={prefetchStrategy} class="grid">
-	<div class="header-bar">
-		<Header {fileTree} bind:open />
-	</div>
-	<Sidebar bind:open {fileTree} />
-	{#if !$navigating}
-		<main in:blur|local id="evidence-content">
-			<div class="content" class:settings-content={!data.isUserPage}>
-				<article class:settings-article={!data.isUserPage}>
-					<slot />
-					<p>&nbsp;</p>
-				</article>
-				{#if data.isUserPage}
-					<aside class="toc">
-						<TableOfContents />
-					</aside>
+<div data-sveltekit-preload-data={prefetchStrategy} class="antialiased text-gray-900">
+	<Header bind:mobileSidebarOpen />
+	<div
+		class="max-w-7xl print:max-w-prose mx-auto print:md:px-0 print:px-0 px-6 sm:px-8 md:px-12 flex justify-start"
+	>
+		<div class="print:hidden">
+			<Sidebar {fileTree} bind:mobileSidebarOpen />
+		</div>
+		<main class="flex-1 overflow-x-hidden md:px-8 print:px-0 print:md:px-0 py-8">
+			<div class="print:hidden">
+				{#if $page.route.id !== '/settings'}
+					<BreadCrumbs {fileTree} />
 				{/if}
 			</div>
+			{#if !$navigating}
+				<!-- <div>
+					<div class="inline-flex h-2 w-2 bg-green-600 rounded-full justify-center items-center">
+						<div class="inline-block h-2 w-2 bg-green-600/30 rounded-full animate-ping" />
+					</div>
+					<span class="px-1 text-xs text-gray-500">Updated 7 hours ago</span>
+				</div> -->
+				<article class="select-text">
+					<slot />
+				</article>
+			{:else}
+				<LoadingSkeleton />
+			{/if}
 		</main>
-	{/if}
+		<div class="print:hidden">
+			<TableOfContents />
+		</div>
+	</div>
 </div>
-{#if !$navigating && dev && data.isUserPage}
-	<QueryStatus />
-{/if}
-
-<style>
-	.grid {
-		display: grid;
-		grid-template-areas:
-			'sidebar header'
-			'sidebar main';
-		grid-template-columns: 18rem 5fr;
-		grid-template-rows: var(--header-height) 1fr;
-		gap: 0 16px;
-		margin: 0 auto;
-		isolation: isolate;
-	}
-
-	main {
-		grid-area: main;
-	}
-
-	div.content {
-		margin: auto;
-		max-width: 100ch;
-		box-sizing: border-box;
-		display: grid;
-		grid-template-columns: 4fr minmax(0, 1fr);
-		gap: 0 5ch;
-		grid-template-areas: 'article toc';
-		justify-items: left;
-	}
-
-	article {
-		max-width: 70ch;
-		min-width: 0;
-		width: 100%;
-		grid-area: article;
-		padding: 0 1.5em 0 1.5em;
-		box-sizing: border-box;
-		user-select: text;
-		-moz-user-select: text;
-		-webkit-user-select: text;
-		-ms-user-select: text;
-	}
-
-	.settings-content {
-		max-width: 100ch !important;
-		grid-template-columns: 1fr !important;
-		grid-template-areas: 'article' !important;
-	}
-
-	.settings-article {
-		max-width: 100ch;
-		min-width: 0;
-		width: 100%;
-		grid-area: article;
-		padding: 0 1.5em 0 1.5em;
-		box-sizing: border-box;
-	}
-	aside.toc {
-		grid-area: toc;
-		padding: 0px;
-	}
-
-	.header-bar {
-		display: flex;
-		justify-content: space-between;
-		position: sticky;
-		z-index: 2;
-		top: 0;
-		background-color: rgba(255, 255, 255, 0.73);
-		-webkit-backdrop-filter: blur(10px) saturate(1.8);
-		backdrop-filter: blur(10px) saturate(1.8);
-	}
-
-	@media (max-width: 1440px) {
-		div.content {
-			grid-template-columns: 1fr;
-			grid-template-areas: 'article';
-		}
-
-		article {
-			max-width: 70ch;
-		}
-
-		.settings-article {
-			max-width: 100ch;
-		}
-
-		.settings-content {
-			max-width: 100ch !important;
-			grid-template-columns: 1fr !important;
-			grid-template-areas: 'article' !important;
-		}
-
-		aside.toc {
-			display: none;
-		}
-	}
-
-	@media (max-width: 850px) {
-		.grid {
-			display: grid;
-			grid-template-areas:
-				'header'
-				'main';
-			grid-template-columns: 1fr;
-			grid-template-rows: var(--header-height) 1fr;
-			margin: 0 auto;
-			isolation: isolate;
-		}
-	}
-
-	@media print {
-		main * {
-			visibility: hidden;
-		}
-
-		.header-bar {
-			visibility: hidden;
-		}
-		article,
-		article * {
-			visibility: visible;
-		}
-		article {
-			position: absolute;
-			left: 0;
-			top: 0;
-		}
-	}
-</style>