+++ sites/example-project/src/pages/settings/+page.svelte
@@ -1,14 +1,20 @@
 <script>
 	export let data;
-	let { settings, customFormattingSettings, gitIgnore } = data;
-	$: ({ settings, customFormattingSettings, gitIgnore } = data);
+	let { settings, customFormattingSettings } = data;
+	$: ({ settings, customFormattingSettings } = data);
 
 	import { dev } from '$app/environment';
+
+	import {
+		VersionControlPanel,
+		DeploySettingsPanel,
+		FormattingSettingsPanel,
+		TelemetrySettingsPanel
+	} from '@evidence-dev/core-components';
 </script>
 
 {#if dev}
 	<!-- eslint-disable no-undef -->
-	<DatabaseSettingsPanel {settings} {gitIgnore} />
 	<VersionControlPanel {settings} />
 	<DeploySettingsPanel {settings} />
 	<FormattingSettingsPanel {settings} {customFormattingSettings} />