+++ packages/core-components/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @evidence-dev/core-components
 
+## 2.0.0-usql.8
+
+### Patch Changes
+
+- Adjust icon usage, dependencies
+
+## 2.0.0-usql.7
+
+### Patch Changes
+
+- Updated dependencies [f051417f]
+  - @evidence-dev/tailwind@1.0.0-usql.1
+
 ## 2.0.0-usql.6
 
 ### Patch Changes