+++ packages/plugin-connector/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @evidence-dev/plugin-connector
 
+## 2.0.0-usql.13
+
+### Patch Changes
+
+- Updated dependencies [69126c94]
+  - @evidence-dev/universal-sql@2.0.0-usql.7
+
 ## 2.0.0-usql.12
 
 ### Patch Changes